lunch_decider/apps/employees/serializers.py
from rest_framework import serializers
from .models import Employee

class EmployeeSerializer(serializers.ModelSerializer):
    class Meta:
        model = Employee
        fields = "__all__"

    def create(self, validated_data):
        # Создание нового экземпляра Employee
        return Employee.objects.create(**validated_data)

    def update(self, instance, validated_data):
        # Обновление существующего экземпляра Employee
        instance.first_name = validated_data.get('first_name', instance.first_name)
        instance.last_name = validated_data.get('last_name', instance.last_name)
        instance.email = validated_data.get('email', instance.email)
        instance.save()
        return instance