lunch_decider/requirements.txt
djangorestframework==3.16.0
jsonpointer==2.1
Markdown==3.8
wheel==0.45.1
asgiref==3.8.1
Django==5.1.3
sqlparse==0.5.2
gunicorn==23.0.0
psycopg2-binary==2.9.10
django-pytest==0.2.0
iniconfig==2.1.0
packaging==25.0
pluggy==1.5.0
pytest==8.3.5
pytest-django==4.11.1